Changed File structure, looking to make the engine neater Working on Terrain Generation, mostly done Height map
looking at restructuring the Engine so that it can be used easier for the rest of the project but it's a bit late to be doing it.
it's mostly to compile the things in the tutorials into something that I can run easily with less dependencies. basically looking to build a program and stop it from being a spider web of dependencies which I make accidentally for the most part